Avoid using ruby's forwardable with private method
Since Ruby 2.4, forwardable now warns whenever a call is forwarded to a
private method.
https://github.com/ruby/ruby/commit/ef13e60fe07766549275c19f3e5112fb7773cf38

In order to avoid filling up our logs, we work around this, by
introducing a regular class method into the calling class.

The fix is actually backported from the following upstream commit:
https://github.com/mongodb/mongo-ruby-driver/commit/53dca7d9381a569549aafee883b5949df22d501a

(server_selector is not present in our release, so no action is needed)