Forwardport ping check before every operation
The latest driver version is unable to handle some writes in the legacy
codepath resulting in Errno::EPIPE: Broken pipe exceptions.

These exceptions arise due to outdated topology state in the short
time window between the monitor thread scans & under the following
conditions:

 - Failover is performed, the previous primary server becomes secondary
   & closes the connection with the driver.
 - The connection from the driver's side changes to CLOSE_WAIT state but
   the driver is clueless about that until the next monitor thread scan.
 - At least 2 writes are performed on the outdated connection.
   On the second write Errno::EPIPE: Broken pipe is raised.

While the same error is possible for reads as well, the driver retries
only the read calls propably due to the inability for bulk-write retries.

In contrary to the latest driver, the previous driver versions are able
to handle those exceptions by sending a ping before each operation.
Pings rescue all exceptions & topology update is performed after a failed ping.

As a workaround for this problem, this ping is re-introduced in the
latest driver version.

NOTE: The modern codepath which is enabled on server versions >=3.6 is
able to handle those exceptions without pinging.